requirements.txt
streamlit
crewai
crewai-tools
python-dotenv
markdown
python-docx
html2text
firecrawl-py